Moved ResizableArray
