+++ src/main/scala/parsley/internal/deepembedding/backend/ResizableArray.scala
@@ -1,10 +1,10 @@
-package parsley.internal
+package parsley.internal.deepembedding.backend
 
 import scala.reflect.ClassTag
 import scala.language.implicitConversions
 
 // This is designed to be a lighter-weight wrapper around Array to make it resizeable
-private [internal] final class ResizableArray[A: ClassTag](initialSize: Int = ResizableArray.InitialSize)
+private [backend] final class ResizableArray[A: ClassTag](initialSize: Int = ResizableArray.InitialSize)
 {
     private [this] var array: Array[A] = new Array(initialSize)
     private [this] var size = 0
@@ -28,13 +28,7 @@ private [internal] final class ResizableArray[A: ClassTag](initialSize: Int = Re
         array = null
         res
     }
-    def toShrunkenArray: Array[A] = if (array.length == size) toArray else {
-        val newArray = new Array[A](size)
-        java.lang.System.arraycopy(array, 0, newArray, 0, size)
-        array = null
-        newArray
-    }
 }
-private [internal] object ResizableArray {
+private [backend] object ResizableArray {
     val InitialSize = 16
 }
