+++ src/main/scala/parsley/internal/machine/instructions/CoreInstrs.scala
@@ -1,7 +1,6 @@
 package parsley.internal.machine.instructions
 
 import parsley.internal.machine.{Context, Good, Recover, Failed, Finished}
-import parsley.internal.ResizableArray
 import parsley.internal.errors.{ErrorItem, Desc}
 import parsley.internal.machine.errors.EmptyError
